Fix sign-conversion warnings

+++ Projects/FileEncryption/src/common/constants.h
@@ -21,7 +21,7 @@ inline constexpr size_t kBlockSize = 16;   /// AES-GCM block size in bytes
 inline constexpr size_t kBuffSize = 4096;  /// Buffer size in blocks
 inline constexpr size_t kIVSize = 12;      /// Initial vector size in bytes
 inline constexpr size_t kTagSize = 16;     /// Authentication tag size in bytes
-inline constexpr int kBuffNum = 2;         /// Number of buffers for swapping
+inline constexpr size_t kBuffNum = 2;      /// Number of buffers for swapping
 
 inline constexpr int64_t kMaxSize = 64ULL * 1024 * 1024 * 1024 - 32;  /// Maximum source file size that can be processed
